Updates go-gnome to more WASM-friendly version. Restructures code.
Nearly all of the changes here were during debugging of the dreaded
"WASM-load-wait-until-click" defect. That defect was in a sub-sub
package, and resolved in go-gnome. I kept the changes here, as they
either added a needed feature (version switch) or genuinely kept
CLI/TUI stuff out of the main GUI/WASM focused file. Welcome to
the first release.